LZW repair alg

+++ src/compression_algorithms/LZW/LZW_core.h
@@ -9,7 +9,7 @@
 #include <stdio.h>
 #include <string.h>
 
-int* lzw_compress_encode(const unsigned char* input,RBTree* dictionary, int* outputSize);
-unsigned char* lzw_decompress_encode(const int* compressed, int compressed_size, RBTree* dictionary);
+int* lzw_compress_encode(const unsigned char* input,RBTree* dictionary, size_t* outputSize);
+unsigned char* lzw_decompress_encode(const int* compressed, size_t compressed_size, RBTree* dictionary);
 
 #endif // TRANSMITR_LZW_CORE_H
